refactor getTopGenres to lib file
